requirements.txt
ollama==0.5.1
PyMuPDF
python-docx